requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-filter==22.1
django-rest-knox==4.2.0
django-sortedm2m==3.1.1
django-sortedone2many==0.2.2
django-unused-media==0.2.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
filelock==3.7.1
idna==3.3
pathlib==1.0.1
Pillow==9.1.1
pipreq==0.4
platformdirs==2.5.2
pycparser==2.21
pytz==2022.1
requests==2.28.1
requires.io==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.14.1
